moved ParentReferenceTests to new location.
